// src/consent/enums/consent.enum.ts
export enum CONSENT_STATUS {
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  REVOKED = 'REVOKED',
  EXPIRED = 'EXPIRED',
}

export enum CONSENT_PURPOSE {
  MARKETING = 'MARKETING',
  ANALYTICS = 'ANALYTICS',
  PERSONALIZATION = 'PERSONALIZATION',
  SHARING = 'SHARING',
  STORAGE = 'STORAGE',
}

export enum CONSENT_ERROR {
  EPOCH_MISMATCH = 'EPOCH_MISMATCH',
  INVALID_PARTNER = 'INVALID_PARTNER',
}